Make `show-pkg` print the names of Compose files for feature flags
