+++ internal/app/forklift/cli/packages.go
@@ -83,7 +83,6 @@ func PrintDeplSpec(indent int, spec core.PkgDeplSpec) {
 	IndentedPrintf(indent, "Deployment:\n")
 	indent++
 
-	// TODO: actually display the app definition
 	IndentedPrintf(indent, "Compose files: ")
 	if len(spec.ComposeFiles) == 0 {
 		fmt.Printf("(none)")
@@ -109,10 +108,23 @@ func PrintFeatureSpecs(indent int, features map[string]core.PkgFeatureSpec) {
 	indent++
 
 	for _, name := range names {
-		if description := features[name].Description; description != "" {
-			IndentedPrintf(indent, "%s: %s\n", name, description)
-			continue
-		}
-		IndentedPrintf(indent, "%s\n", name)
+		PrintFeatureSpec(indent, name, features[name])
+	}
+}
+
+func PrintFeatureSpec(indent int, name string, spec core.PkgFeatureSpec) {
+	IndentedPrintf(indent, "%s:\n", name)
+	indent++
+
+	IndentedPrintf(indent, "Description: %s\n", spec.Description)
+
+	IndentedPrintf(indent, "Compose files: ")
+	if len(spec.ComposeFiles) == 0 {
+		fmt.Printf("(none)")
+		return
+	}
+	fmt.Println()
+	for _, file := range spec.ComposeFiles {
+		BulletedPrintln(indent+1, file)
 	}
 }
