+++ internal/app/forklift/cli/pallets-deployments.go
@@ -2,12 +2,14 @@ package cli
 
 import (
 	"fmt"
+	"path"
 	"sort"
 
 	dct "github.com/compose-spec/compose-go/types"
 	"github.com/pkg/errors"
 
 	"github.com/PlanktoScope/forklift/internal/app/forklift"
+	"github.com/PlanktoScope/forklift/internal/clients/docker"
 	"github.com/PlanktoScope/forklift/pkg/core"
 )
 
@@ -131,4 +133,18 @@ func printDockerAppServices(indent int, services []dct.ServiceConfig) {
 	for _, service := range services {
 		IndentedPrintf(indent, "%s: %s\n", service.Name, service.Image)
 	}
+}
+
+func loadAppDefinition(pkg *core.FSPkg) (*dct.Project, error) {
+	appDef, err := docker.LoadAppDefinition(
+		pkg.FS, path.Base(pkg.Path()), pkg.Def.Deployment.ComposeFiles, nil,
+	)
+	// TODO: also load the docker compose files for all features
+	if err != nil {
+		return nil, errors.Wrapf(
+			err, "couldn't load Docker Compose app definition for a basic deployment of %s",
+			pkg.FS.Path(),
+		)
+	}
+	return appDef, nil
 }
